requirements.txt
addict==2.4.0
av==12.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.1.1
cycler==0.12.1
einops==0.8.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
future==1.0.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.3
requests==2.32.4
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
six==1.17.0
sympy==1.13.3
tifffile==2023.7.10
timm==1.0.22
tomli==2.3.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
yapf==0.43.0
zipp==3.20.2
